Cleanup BGP Neighbor type conversions
Makes a bunch of the version-suffixed API endpoint handlers into default
methods. Updates the calling convention for all the different versioned
endpoint handlers such that they each call the next latest version and
convert via .into(). Also updates the bgp_apply path to properly walk
the conversion path v1 -> v4 -> v8 instead of skipping v4. Also updates
the bgp_apply group removal test to call do_bgp_apply with latest rather
than using a conversion chain.
