+++ mg-api-types/versions/src/bgp_src_addr/bgp/config.rs
@@ -191,36 +191,6 @@ impl From<v1::bgp::config::BgpPeerConfig> for BgpPeerConfig {
     }
 }
 
-impl From<v1::bgp::config::ApplyRequest> for ApplyRequest {
-    fn from(req: v1::bgp::config::ApplyRequest) -> Self {
-        // v1 is schema-stabilized; new schema fields cannot land here.
-        // If this destructure stops compiling, either the addition is
-        // a runtime-only field (#[serde(skip)] / #[schemars(skip)] —
-        // add it to the destructure with `_:`) or the v1 contract has
-        // been violated upstream.
-        let v1::bgp::config::ApplyRequest {
-            asn,
-            originate,
-            checker,
-            shaper,
-            peers,
-        } = req;
-        Self {
-            asn,
-            originate: originate.iter().map(|p| Prefix::V4(*p)).collect(),
-            checker,
-            shaper,
-            peers: peers
-                .into_iter()
-                .map(|(k, v)| {
-                    (k, v.into_iter().map(BgpPeerConfig::from).collect())
-                })
-                .collect(),
-            unnumbered_peers: HashMap::default(),
-        }
-    }
-}
-
 // ----- v4 (mp_bgp, frozen) <-> v8 BgpPeerParameters -----
 
 impl From<BgpPeerParameters> for v4::bgp::config::BgpPeerParameters {
@@ -541,116 +511,3 @@ impl From<v5::bgp::config::UnnumberedNeighbor> for UnnumberedNeighbor {
     }
 }
 
-// ----- v8 -> v1 (initial, frozen) downgrades for /v1 endpoint responses -----
-//
-// v1 lacks per-AF policies, per-AF nexthops, and source-address binding.
-// These conversions fold the per-AF allow_import4/6 and allow_export4/6
-// policies back into v1's single allow_import / allow_export fields via
-// `ImportExportPolicy::from_per_af_policies`, and drop the fields that
-// v1 cannot represent.
-
-impl From<BgpPeerParameters> for v1::bgp::config::BgpPeerParameters {
-    fn from(p: BgpPeerParameters) -> Self {
-        let BgpPeerParameters {
-            hold_time,
-            idle_hold_time,
-            delay_open,
-            connect_retry,
-            keepalive,
-            resolution,
-            passive,
-            remote_asn,
-            min_ttl,
-            md5_auth_key,
-            multi_exit_discriminator,
-            communities,
-            local_pref,
-            enforce_first_as,
-            vlan_id,
-            ipv4_unicast,
-            ipv6_unicast,
-            // v1 has no jitter, no deterministic collision resolution,
-            // and no source-address binding — these are all dropped.
-            deterministic_collision_resolution: _,
-            idle_hold_jitter: _,
-            connect_retry_jitter: _,
-            src_addr: _,
-            src_port: _,
-        } = p;
-        let (import4, export4) = match ipv4_unicast {
-            Some(Ipv4UnicastConfig {
-                import_policy,
-                export_policy,
-                nexthop: _,
-            }) => (import_policy, export_policy),
-            None => Default::default(),
-        };
-        let (import6, export6) = match ipv6_unicast {
-            Some(Ipv6UnicastConfig {
-                import_policy,
-                export_policy,
-                nexthop: _,
-            }) => (import_policy, export_policy),
-            None => Default::default(),
-        };
-        Self {
-            hold_time,
-            idle_hold_time,
-            delay_open,
-            connect_retry,
-            keepalive,
-            resolution,
-            passive,
-            remote_asn,
-            min_ttl,
-            md5_auth_key,
-            multi_exit_discriminator,
-            communities,
-            local_pref,
-            enforce_first_as,
-            allow_import:
-                v1::bgp::policy::ImportExportPolicy::from_per_af_policies(
-                    &import4, &import6,
-                ),
-            allow_export:
-                v1::bgp::policy::ImportExportPolicy::from_per_af_policies(
-                    &export4, &export6,
-                ),
-            vlan_id,
-        }
-    }
-}
-
-impl From<BgpPeerConfig> for v1::bgp::config::BgpPeerConfig {
-    fn from(cfg: BgpPeerConfig) -> Self {
-        let BgpPeerConfig {
-            host,
-            name,
-            parameters,
-        } = cfg;
-        Self {
-            host,
-            name,
-            parameters: v1::bgp::config::BgpPeerParameters::from(parameters),
-        }
-    }
-}
-
-impl From<Neighbor> for v1::bgp::config::Neighbor {
-    fn from(n: Neighbor) -> Self {
-        let Neighbor {
-            asn,
-            name,
-            group,
-            host,
-            parameters,
-        } = n;
-        Self {
-            asn,
-            name,
-            group,
-            host,
-            parameters: v1::bgp::config::BgpPeerParameters::from(parameters),
-        }
-    }
-}
